Add YVBKeyLogger class and base methods
This class allows you to listen to the key up and key down events from
the system. The interface provides a block that will let the user pass
what the action will be capturing the event.

If no handler is passed all the output will be logged to standard output
with a predefined format.