Update class to capture metadata correctly
The callback was ignoring whether the key was a kCGKeyDown or a kCGKeyUp
event hence making it impossible to differentiate between these and
making each key-pressing event repeated.

Additionally a convenience method has been added
(requestEnableAccessibility) to check whether or not the current process
is capable of capturing keystrokes of the following keys a-z. A-Z, 0-9.