Change property to be copy and add class method
The new class method +requestEnableAccessibility allows the user to
request the application to be added to the whitelist of programs that
can read the keys that are being pressed. It's important to have this as
a class method because the logger will be pretty much useless if there's
no authorization.

The property for the block was changed to make it copy so that ARC can
deal with all the memory management.