Add main labels to the program's window
Four labels have been added these labels are going to be presenting the
information that's retrieved from the user-triggered events.

Additionally all the code that was living inside the block has now been
moved to live outside it (when possible) so that there's less overhead
when executing queries using the database object.

As it is the program is now capable of registring all they key-pressed
events and add them to the sqlite database as well as display this
information on screen.