Add fix for system time changes
This feature is extremely tricky, requires the object to invalidate the
timer and then re-scheduling it. Note that we have to invalidate it in a
very odd way, we have do it by just setting the value to nil and not
calling the method invalidate otherwise the application will crash in a
very mysterious and obscure way.

The object now listens to a notification that gets called when the
system clock is changed i. e. if the user changed the clock of the
system, this notification would update the timer such that it waited the
appropriate amount of time.

The timer property of this executor is now readonly and not just a
private property.