Check for events that disable the KeyLogger
When a callback taks a long time one of the following two events can be
returned as part of the callback:
    - kCGEventTapDisabledByTimeout
    - kCGEventTapDisabledByUserInput

Any of which preceeds the operating system from removing the callback
from the CFRunLoop.

The strategy that we've decided to take is observer pattern using the
default notification center to let the main thread know that the
KeyLogger has halted it's execution.

The notification in cuestion is YVBKeyLoggerPerishedNotification.