ENH: Add frontmostApplication to the database
Update data manager to the an application name along with the keystroke
and other information in the database. This required the tests to be
updated as well, along with the database files.

Fixing any existing database should be a matter of:

$ sqlite3 ~/Library/Application\ Support/Keystats/keystrokes
SQLite version 3.8.4.3 2014-04-03 16:53:12
Enter ".help" for usage hints.
sqlite> ALTER TABLE keystrokes ADD COLUMN application varchar(32);