MAINT: Change daily executor internals
The YVBDailyExecutor class no longer relies on a single NSTimer that is
meant to be executed at midnight, instead it now checks every five
minutes and compares whether or not the current date (from which it was
initialized) has changed, if this is true then it updates the current
date and repeats.

I have also removed the subscription to
NSSystemClockDidChangeNotification because this made the system execute
the handler block whenever the computer went to sleep and whenever the
computer awoke.

The main reason to do this and not using NSTimer and subscribing to the
sleep/awake notifications to account for the time changes is because we
would still need a notification when the system clock changed however,
this notification got triggered every time there was a sleep/awake event
which was pointless.