BUG: Fix bug where new day would not be rendered
When the keystrokes per day plot was updated and the new day comming in
didn't have any keystrokes, the value would not be plotted and the data
would be truncated to only fit the days that had already been seen.

To accommodate this change the __datesData private variable was changed
to be an NSMutableArray.