BUG: Fix bug with x-tick positions
We had to move from a fixed-interval labeling policy to a
location-provided one. The main reason is because the location of each
bar varies depending on what time-zone (or whatever other voodoo) the
data was collected at, so to be perfectly accurate: ticks are now
located at the same position where the bars are.

I first encountered this problem when we went through daylight savings a
few days ago.